DONE_PyTest_Example!!!/15. API-framework-brusnika - Allure/requirements.txt
allure-combine==1.0.11
allure-pytest==2.13.5
allure-python-commons==2.13.5
pytest==8.2.2
pytest-timeout==2.3.1
pytest-xdist==3.6.1
requests==2.31.0
